+++ tests/src/test/scala/ldbc/tests/TableQuerySelectConnectionTest.scala
@@ -6,8 +6,6 @@
 
 package ldbc.tests
 
-import java.time.LocalDate
-
 import com.mysql.cj.jdbc.MysqlDataSource
 
 import cats.data.Kleisli
@@ -20,12 +18,13 @@ import org.typelevel.otel4s.trace.Tracer
 import munit.*
 
 import ldbc.core.*
-import ldbc.core.model.*
 import ldbc.sql.*
 import ldbc.sql.logging.LogHandler
 import ldbc.query.builder.TableQuery
 import ldbc.connector.SSL
 import ldbc.dsl.io.*
+
+import ldbc.tests.model.*
 
 class LdbcTableQuerySelectConnectionTest extends TableQuerySelectConnectionTest:
 
@@ -62,134 +61,6 @@ trait TableQuerySelectConnectionTest extends CatsEffectSuite:
 
   def prefix:     "jdbc" | "ldbc"
   def connection: Resource[IO, Connection[IO]]
-
-  case class Country(
-    code:           String,
-    name:           String,
-    continent:      Country.Continent,
-    region:         String,
-    surfaceArea:    BigDecimal,
-    indepYear:      Option[Short],
-    population:     Int,
-    lifeExpectancy: Option[BigDecimal],
-    gnp:            Option[BigDecimal],
-    gnpOld:         Option[BigDecimal],
-    localName:      String,
-    governmentForm: String,
-    headOfState:    Option[String],
-    capital:        Option[Int],
-    code2:          String
-  )
-
-  object Country:
-
-    enum Continent(val value: String) extends Enum:
-      case Asia          extends Continent("Asia")
-      case Europe        extends Continent("Europe")
-      case North_America extends Continent("North America")
-      case Africa        extends Continent("Africa")
-      case Oceania       extends Continent("Oceania")
-      case Antarctica    extends Continent("Antarctica")
-      case South_America extends Continent("South America")
-
-      override def toString: String = value
-
-    object Continent extends EnumDataType[Continent]
-
-    given Parameter[IO, Continent] with
-      override def bind(statement: PreparedStatement[IO], index: Int, value: Continent): IO[Unit] =
-        statement.setString(index, value.toString)
-
-    given ResultSetReader[IO, Continent] =
-      ResultSetReader.mapping[IO, String, Continent](str => Continent.valueOf(str.replace(" ", "_")))
-
-    val table: Table[Country] = Table[Country]("country")(
-      column("Code", CHAR(3).DEFAULT(""), PRIMARY_KEY),
-      column("Name", CHAR(52).DEFAULT("")),
-      column("Continent", ENUM(using Continent).DEFAULT(Continent.Asia)),
-      column("Region", CHAR(26).DEFAULT("")),
-      column("SurfaceArea", DECIMAL(10, 2).DEFAULT(0.00)),
-      column("IndepYear", SMALLINT.DEFAULT(None)),
-      column("Population", INT.DEFAULT(0)),
-      column("LifeExpectancy", DECIMAL(3, 1).DEFAULT(None)),
-      column("GNP", DECIMAL(10, 2).DEFAULT(None)),
-      column("GNPOld", DECIMAL(10, 2).DEFAULT(None)),
-      column("LocalName", CHAR(45).DEFAULT("")),
-      column("GovernmentForm", CHAR(45).DEFAULT("")),
-      column("HeadOfState", CHAR(60).DEFAULT(None)),
-      column("Capital", INT.DEFAULT(None)),
-      column("Code2", CHAR(2).DEFAULT(""))
-    )
-
-  case class City(
-    id:          Int,
-    name:        String,
-    countryCode: String,
-    district:    String,
-    population:  Int
-  )
-
-  object City:
-
-    val table: Table[City] = Table[City]("city")(
-      column("ID", INT, AUTO_INCREMENT, PRIMARY_KEY),
-      column("Name", CHAR(35).DEFAULT("")),
-      column("CountryCode", CHAR(3).DEFAULT("")),
-      column("District", CHAR(20).DEFAULT("")),
-      column("Population", INT.DEFAULT(0))
-    )
-      .keySet(v => INDEX_KEY(v.countryCode))
-      .keySet(v => CONSTRAINT("city_ibfk_1", FOREIGN_KEY(v.countryCode, REFERENCE(Country.table, Country.table.code))))
-
-  case class CountryLanguage(
-    countryCode: String,
-    language:    String,
-    isOfficial:  CountryLanguage.IsOfficial,
-    percentage:  BigDecimal
-  )
-
-  object CountryLanguage:
-
-    enum IsOfficial extends Enum:
-      case T, F
-
-    object IsOfficial extends EnumDataType[IsOfficial]
-
-    given Parameter[IO, IsOfficial] with
-      override def bind(statement: PreparedStatement[IO], index: Int, value: IsOfficial): IO[Unit] =
-        statement.setString(index, value.toString)
-
-    given ResultSetReader[IO, IsOfficial] =
-      ResultSetReader.mapping[IO, String, IsOfficial](str => IsOfficial.valueOf(str))
-
-    val table: Table[CountryLanguage] = Table[CountryLanguage]("countrylanguage")(
-      column("CountryCode", CHAR(3).DEFAULT("")),
-      column("Language", CHAR(30).DEFAULT("")),
-      column("IsOfficial", ENUM(using IsOfficial).DEFAULT(IsOfficial.F)),
-      column("Percentage", DECIMAL(4, 1).DEFAULT(0.0))
-    )
-      .keySet(v => PRIMARY_KEY(v.countryCode, v.language))
-      .keySet(v => INDEX_KEY(v.countryCode))
-      .keySet(v =>
-        CONSTRAINT("countryLanguage_ibfk_1", FOREIGN_KEY(v.countryCode, REFERENCE(Country.table, Country.table.code)))
-      )
-
-  case class GovernmentOffice(
-    id:                Int,
-    cityId:            Int,
-    name:              String,
-    establishmentDate: Option[LocalDate]
-  )
-
-  object GovernmentOffice:
-
-    val table: Table[GovernmentOffice] = Table[GovernmentOffice]("government_office")(
-      column("ID", INT, AUTO_INCREMENT, PRIMARY_KEY),
-      column("CityID", INT),
-      column("Name", CHAR(35).DEFAULT("")),
-      column("EstablishmentDate", DATE)
-    )
-      .keySet(v => CONSTRAINT("government_office_ibfk_1", FOREIGN_KEY(v.cityId, REFERENCE(City.table, City.table.id))))
 
   private final val country          = TableQuery[IO, Country](Country.table)
   private final val city             = TableQuery[IO, City](City.table)
