+++ tests/src/test/scala/ldbc/tests/model/City.scala
@@ -1,0 +1,29 @@
+/**
+ * Copyright (c) 2023-2024 by Takahiko Tominaga
+ * This software is licensed under the MIT License (MIT).
+ * For more information see LICENSE or https://opensource.org/licenses/MIT
+ */
+
+package ldbc.tests.model
+
+import ldbc.core.*
+
+case class City(
+                 id: Int,
+                 name: String,
+                 countryCode: String,
+                 district: String,
+                 population: Int
+               )
+
+object City:
+
+  val table: Table[City] = Table[City]("city")(
+    column("ID", INT, AUTO_INCREMENT, PRIMARY_KEY),
+    column("Name", CHAR(35).DEFAULT("")),
+    column("CountryCode", CHAR(3).DEFAULT("")),
+    column("District", CHAR(20).DEFAULT("")),
+    column("Population", INT.DEFAULT(0))
+  )
+    .keySet(v => INDEX_KEY(v.countryCode))
+    .keySet(v => CONSTRAINT("city_ibfk_1", FOREIGN_KEY(v.countryCode, REFERENCE(Country.table, Country.table.code))))
